fix: Updated LazyDatasset class.

+++ test_updated.py
@@ -6,6 +6,7 @@ from glob import glob
 import multiprocessing
 from tqdm import tqdm
 from functools import lru_cache
+import json
 
 # Class definitions
 class SmallDataset(InMemoryDataset):
@@ -141,7 +142,9 @@ def load_batch(path, weights_only=True):
     return list(torch.load(path, weights_only=weights_only))
 
 class LazyDataset(Dataset):
-    def __init__(self, root, transform=None, pre_transform=None, pre_filter=None, datalist=None, num_workers=8, chunk_size=100, pickle_protocol=5, clean_keys=('is_data', 'rec_indices'), batch_size=1000, use_cache=False, drop_last=False, weights_only=False):
+
+    def __init__(self, root, transform=None, pre_transform=None, pre_filter=None, datalist=None, num_workers=0, chunk_size=100, pickle_protocol=5, clean_keys=('is_data', 'rec_indices'), batch_size=100000, drop_last=False, weights_only=False, recreate=False):
+        self.metadata_file_name = 'metadata.json'
         self.datalist = datalist
         self.root = root
         self.num_workers = num_workers
@@ -149,20 +152,57 @@ class LazyDataset(Dataset):
         self.chunk_size = chunk_size
         self.clean_keys = clean_keys
         self.batch_size = batch_size
-        self._loaded_batch = None
-        self._loaded_batch_idx = -1
         self.drop_last = drop_last
         self.num_batches =  0
         if self.datalist is not None and len(self.datalist)>0 and self.batch_size>0:
             self.num_batches = len(self.datalist)//self.batch_size+(1 if len(self.datalist)%self.batch_size>0 and not self.drop_last else 0)
-        self.use_cache = use_cache
+        self.size = len(datalist) if datalist is not None else 0
         self.weights_only = weights_only
+        self.process_batch_start_idx = 0
+
+        # Check if metadata already exists
+        metadata_path = os.path.join(self.root,self.metadata_file_name)
+        metadata = {"size":self.size, "batch_size": self.batch_size, "num_batches":self.num_batches}
+        if not recreate and osp.exists(metadata_path):
+
+            # First try opening file and reading metadata
+            try:
+                with open(metadata_path, "r") as f:
+
+                    # Read metadata from previously saved dataset
+                    metadata = json.load(f)
+                    self.size += metadata["size"] #NOTE: Ordering matters here.
+                    self.batch_size = metadata["batch_size"]
+                    self.num_batches = self.size//self.batch_size+(1 if self.size%self.batch_size>0 and not self.drop_last else 0)
+                    self.process_batch_start_idx = metadata["num_batches"]-1
+
+                    # Reset metadata for when you write it below
+                    metadata = {"size":self.size, "batch_size": self.batch_size, "num_batches":self.num_batches}
+
+            except FileNotFoundError as e:
+                print("Metadata file not found, dataset is corrupted!")
+                raise e
+
+            # Load last batch and check if length is same as batch size.
+            # If it is not, then prepend this batch to self.datalist so that everything gets saved correctly.
+            # It is important to use torch.load and not load_batch here so that you don't end up loading
+            # a partial batch later on!
+            _loaded_batch = torch.load(os.path.join(self.processed_dir, self.processed_file_names[self.process_batch_start_idx]), weights_only=self.weights_only)
+            if len(_loaded_batch)!=self.batch_size:
+                if self.datalist is None: self.datalist = []
+                self.datalist = [*_loaded_batch, *self.datalist]
+
+        # (Re)create the metadata file
+        os.makedirs(self.root, exist_ok=True)
+        with open(metadata_path, "w") as f:
+            json.dump(metadata, f, indent=2)
+        
         super().__init__(root, transform, pre_transform, pre_filter)
 
     @property
     def raw_file_names(self):
         if self.datalist is not None and len(self.datalist)>0:
-            return [f'data{i}.pt' for i in range(self.num_batches)]
+            return [f'data{i}.pt' for i in range(self.size)]
         else:
             return [os.path.basename(path) for path in glob(os.path.join(self.raw_dir, 'data*.pt'))]
 
@@ -174,8 +214,6 @@ class LazyDataset(Dataset):
             return [os.path.basename(path) for path in glob(os.path.join(self.processed_dir, 'data*.pt'))]
 
     def clean_data(self,data):
-
-        # Create a new graph and remove undesired attributes
         cleaned_data = Data()
         for key in data.keys():
             if key in (self.clean_keys): continue
@@ -186,24 +224,20 @@ class LazyDataset(Dataset):
 
     def save_graph_batch(self,idx):
 
-        # Set batch indices
         min_idx = idx*self.batch_size
-        max_idx = min((idx+1)*self.batch_size, len(self.datalist)-1)
+        max_idx = min((idx+1)*self.batch_size, len(self.datalist))
 
-        # Select batch
-        data = self.datalist[min_idx:max_idx+1]
+        data = self.datalist[min_idx:max_idx]
 
-        # Apply filters and transforms
         if self.pre_filter is not None:
             data = [d for d in data if not self.pre_filter(data)]
+
         if self.pre_transform is not None:
             data = [self.pre_transform(d) for d in data]
 
-        # Clean batch
         data = [self.clean_data(d) for d in data]
 
-        # Save batch
-        torch.save(data, osp.join(self.processed_dir, self.processed_file_names[idx]), pickle_protocol=self.pickle_protocol)
+        torch.save(data, osp.join(self.processed_dir, self.processed_file_names[idx+self.process_batch_start_idx]), pickle_protocol=self.pickle_protocol)
 
     def process(self):
 
@@ -211,36 +245,38 @@ class LazyDataset(Dataset):
         if self.datalist is None or len(self.datalist)==0:
             return
 
-        # Save batches of graphs in several processes
-        with multiprocessing.Pool(processes=min(self.num_batches, self.num_workers)) as pool:
-            try:
-                list(tqdm(pool.imap_unordered(self.save_graph_batch, range(self.num_batches), self.chunk_size), total=self.num_batches))
-            except KeyBoardInterrupt as e:
-                print("Caught KeyBoardInterrupt, terminating workers")
-                pool.terminate()
-                pool.join()
-                print(e)
-            else:
-                pool.close()
-                pool.join()
+        local_num_batches = self.num_batches - self.process_batch_start_idx
+
+        if self.num_workers<=0:
+            tqdm([self.save_graph_batch(idx) for idx in range(local_num_batches)])
+        else:
+            with multiprocessing.Pool(processes=min(local_num_batches, self.num_workers)) as pool:
+                try:
+                    list(tqdm(pool.imap_unordered(self.save_graph_batch, range(local_num_batches), self.chunk_size), total=local_num_batches))
+                except KeyboardInterrupt as e:
+                    print("Caught KeyBoardInterrupt, terminating workers")
+                    pool.terminate()
+                    pool.join()
+                    print(e)
+                else:
+                    pool.close()
+                    pool.join()
 
     def len(self):
         return len(self.processed_file_names)*self.batch_size
 
     def get(self, idx):
 
+        # Check index
+        if idx>=self.size:
+            raise IndexError(f'Index {idx} out of range for dataset of length {len(self)}')
+
         # Get indices
         batch_idx = idx // self.batch_size
         within_idx = idx % self.batch_size
 
-        # Load batch
-        if batch_idx != self._loaded_batch_idx or not self.use_cache:
-            # self._loaded_batch = torch.load(os.path.join(self.processed_dir, self.processed_file_names[batch_idx]), weights_only=self.weights_only)
-            # self._loaded_batch_idx = batch_idx
-            # data = self._loaded_batch[within_idx]
-            _loaded_batch = load_batch(os.path.join(self.processed_dir, self.processed_file_names[batch_idx]), weights_only=self.weights_only)
-
-        # Load and transform data from batch
+        # Load batch data and select at index
+        _loaded_batch = load_batch(os.path.join(self.processed_dir, self.processed_file_names[batch_idx]), weights_only=self.weights_only)
         data = _loaded_batch[within_idx]
         return data
